Rename convindex.py to semrefindex.py
